Cleanup based on code review
