+++ spring-cloud-data-module-deployers/spring-cloud-data-module-deployer-lattice/src/main/java/org/springframework/cloud/data/module/deployer/lattice/LrpModuleDeployer.java
@@ -50,7 +50,7 @@ public class LrpModuleDeployer implements ModuleDeployer {
 
 	private final ReceptorClient receptorClient = new ReceptorClient();
 
-	private final ReceptorProcessStatusMapper receptorProcessStatusMapper = new ReceptorProcessStatusMapper();
+	private final StatusMapper receptorProcessStatusMapper = new StatusMapper();
 
 	@Override
 	public ModuleDeploymentId deploy(ModuleDeploymentRequest request) {
@@ -132,4 +132,38 @@ public class LrpModuleDeployer implements ModuleDeployer {
 	public Map<ModuleDeploymentId, ModuleStatus> status() {
 		throw new UnsupportedOperationException();
 	}
+
+	public static class StatusMapper {
+
+		public ModuleStatus.State map(ActualLRPResponse actualLRPResponse) {
+			ModuleStatus.State state;
+
+			switch (actualLRPResponse.getState()) {
+				case "RUNNING":
+					state = ModuleStatus.State.deployed;
+					break;
+				case "UNCLAIMED":
+					// see description of UNCLAIMED here: https://github.com/cloudfoundry-incubator/receptor/blob/master/doc/lrps.md
+					if (StringUtils.hasText(actualLRPResponse.getPlacementError())) {
+						state = ModuleStatus.State.failed;
+					}
+					else {
+						state = ModuleStatus.State.deploying;
+					}
+					break;
+				case "CLAIMED":
+					state = ModuleStatus.State.deploying;
+					break;
+				case "CRASHED":
+					state = ModuleStatus.State.failed;
+					break;
+				default:
+					state = ModuleStatus.State.unknown;
+			}
+
+			return state;
+
+		}
+	}
+
 }
