+++ spring-cloud-data-module-deployers/spring-cloud-data-module-deployer-lattice/src/main/java/org/springframework/cloud/data/module/deployer/lattice/TaskModuleDeployer.java
@@ -48,7 +48,7 @@ public class TaskModuleDeployer implements ModuleDeployer {
 
 	private final ReceptorClient receptorClient = new ReceptorClient();
 
-	private final ReceptorTaskStatusMapper statusMapper = new ReceptorTaskStatusMapper();
+	private final StatusMapper statusMapper = new StatusMapper();
 
 	@Override
 	public ModuleDeploymentId deploy(ModuleDeploymentRequest request) {
@@ -133,4 +133,38 @@ public class TaskModuleDeployer implements ModuleDeployer {
 	public Map<ModuleDeploymentId, ModuleStatus> status() {
 		throw new UnsupportedOperationException();
 	}
+
+	public static class StatusMapper {
+
+		public ModuleStatus.State map(TaskResponse taskResponse) {
+			ModuleStatus.State state;
+
+			switch (taskResponse.getState()) {
+				case "PENDING":
+					state = ModuleStatus.State.deploying;
+					break;
+				case "CLAIMED":
+					state = ModuleStatus.State.deploying;
+					break;
+				case "RUNNING":
+					//TODO: Add support for canceling
+					state = ModuleStatus.State.deployed;
+					break;
+				case "COMPLETED":
+					//TODO: Add support for canceled
+					if(taskResponse.isFailed()) {
+						state = ModuleStatus.State.failed;
+					}
+					else {
+						state = ModuleStatus.State.complete;
+					}
+					break;
+				default:
+					state = ModuleStatus.State.unknown;
+			}
+
+			return state;
+		}
+	}
+
 }
